NOTE: Requires Visual Studio 2008 SP1 / .NET 3.5 SP1 due to use of [DataContract(IsReference=true)] attribute, which is not available in .NET 3.5 Feature: Serialization will now use entity reference notation, rather than duplicating entities (see DataContract attribute "IsReference" property in MSND documentation).  This also means that serialization string is smaller as no data is duplicated, and it also means that when de-serializing entity references will be resolved correctly, rather than making copies of objects. Bug Fix: Fixed deletion issue where entities with multiple parents caused exception when deletion logic attempted to delete the object twice.
git-svn-id: https://linq2sqleb.svn.codeplex.com/svn@17306 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ Source/LINQEntityBase/LINQEntityBaseExampleData/LINQEntityBase.cs
@@ -45,7 +45,7 @@ namespace LINQEntityBaseExampleData
     /// e.g. EntityBase="LINQEntityBaseExample.LINQEntityBase"
     /// 
     /// </summary>    
-    [DataContract()]
+    [DataContract(IsReference=true)]
     [KnownType("GetKnownTypes")]
     public abstract class LINQEntityBase
     {
@@ -518,17 +518,27 @@ namespace LINQEntityBaseExampleData
                         // Cascade delete children and then this object
                         foreach (LINQEntityBase toDelete in entityTreeReversed)
                         {
-                            toDelete.SetAsDeleteOnSubmit();
-                            targetDataContext.GetTable(toDelete.GetEntityType()).Attach(toDelete);
-                            targetDataContext.GetTable(toDelete.GetEntityType()).DeleteOnSubmit(toDelete);
+                            // Before we try and delete, make sure the entity hasn't been marked to be deleted already
+                            // through another relationship linkng this entity in the same sub-tree that is being deleted.
+                            if (!entitiesDeleted.Contains(toDelete))
+                            {
+                                // Mark for deletion
+                                toDelete.SetAsDeleteOnSubmit();
+                                targetDataContext.GetTable(toDelete.GetEntityType()).Attach(toDelete);
+                                targetDataContext.GetTable(toDelete.GetEntityType()).DeleteOnSubmit(toDelete);
+
+                                //add deleted entity to a list to make sure we don't delete them twice.
+                                entitiesDeleted.Add(toDelete);
+                            }
                         }
-                        //add these to a list to make sure we don't attach them twice.
-                        entitiesDeleted.AddRange(entityTreeReversed);
                     }
                     else
                     {
+                        // Mark for deletion
                         targetDataContext.GetTable(entity.GetEntityType()).Attach(entity);
                         targetDataContext.GetTable(entity.GetEntityType()).DeleteOnSubmit(entity);
+                        
+                        //add deleted entity to a list to make sure we don't delete them twice.
                         entitiesDeleted.Add(entity);
                     }
